🚚 Rename: 채팅방 목록 더미데이터 파일명 변경 (#154)
